+++ templates/servicio/reporte_servicios.php
@@ -101,7 +101,7 @@
                     <?php if ($_SESSION['tipo'] != 2) { ?>
                         <nav class="card--title__menu">
                             <a href="form_modificar_servicio.php?servicio=<?php echo $servicio['clave_servicio'] ?>" class="card--title__item"><i class="icon-edit"></i></a>
-                            <a href="eliminar_servicio.php?servicio=<?php echo $servicio['clave_servicio'] ?>" class="card--title__item"><i class="icon-delete"></i></a>
+                            <a href="eliminar_servicio.php?servicio=<?php echo $servicio['clave_servicio'] ?>" onclick="return Confirmar_Eliminar()" class="card--title__item"><i class="icon-delete"></i></a>
                         </nav>
                     <?php } ?>
                 </div>
